third_party/WebKit/LayoutTests/fast/dom/shadow/event-path-multiple-shadow-roots-2-expected.txt


click #B

event.path on node #B (target: #B)
#B, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 6

event.path on node #H (target: #B)
#B, #H, #SR-MIDDLE, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 8

event.path on node #SR-MIDDLE (target: #B)
#B, #H, #SR-MIDDLE, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 8

event.path on node #F (target: #B)
#B, #H, #SR-MIDDLE, #F, #SR-YOUNGEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 10

event.path on node #SR-YOUNGEST (target: #B)
#B, #H, #SR-MIDDLE, #F, #SR-YOUNGEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 10

event.path on node #A (target: #B)
#B, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 6

click #A/E

event.path on node #E (target: #E)
#E, #SR-OLDEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 7

event.path on node #SR-OLDEST (target: #E)
#E, #SR-OLDEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 7

event.path on node #A (target: #A)
#A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 5

click #A//I

event.path on node #I (target: #I)
#I, #SR-MIDDLE, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 7

event.path on node #SR-MIDDLE (target: #I)
#I, #SR-MIDDLE, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 7

event.path on node #F (target: #I)
#I, #SR-MIDDLE, #F, #SR-YOUNGEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 9

event.path on node #SR-YOUNGEST (target: #I)
#I, #SR-MIDDLE, #F, #SR-YOUNGEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 9

event.path on node #A (target: #A)
#A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 5

click #A///G

event.path on node #G (target: #G)
#G, #SR-YOUNGEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 7

event.path on node #SR-YOUNGEST (target: #G)
#G, #SR-YOUNGEST, #A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 7

event.path on node #A (target: #A)
#A, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], length: 5
PASS successfullyParsed is true

TEST COMPLETE
